Fixed inertial scroll not being cancelled when entering overscroll
The problem here was that if inertial scroll was so strong that it caused an overscroll, and the overscroll came to a rest before the inertial scroll had naturally decelerated, the inertial scroll would kick back in when overscroll came to a rest. This has been fixed by clearing the inertial scroll on tick if it is not valid for it to occur.

Also fixed an issue where scroll samples would not be added when a touch-initated scroll was not moving. This could mean that a fast scroll followed by holding the scroll still would result in the inertial scroll kicking in with a fast velocity when the touch was released. This has been fixed by adding scroll samples in the tick function rather than on mouse/touch moved.

Reviewed by Nick Atamas

[CL 2274268 by Andrew Rodham in Main branch]