Graceful handling of PlatformInitMultimedia failure.
Headless programs (servers) may want to throw message boxes. Without this change, they would be failing early attempting to initialize SDL for that purpose (this actually happened when a plugin was missing from a server build), and the error message about SDL initialization failure is unhelpful to pinpoint the actual problem.

[CL 2277213 by Dmitry Rekman in Main branch]