Duplicate-named Blackboard entries can no longer be created.
Added checks for duplicate names when renaming entries in the Blackboard editor.
Added check when creating new keys so that we dont create duplicates at first.
Disabled name editing in the details panel, as we shoud only be editing the names in the list.

TTP# 343597 - BlackBoard editor allows multiple keys of the same name [modified]

[CL 2277513 by Thomas Sarkanen in Main branch]