Fixed measure tool calculating the wrong size
The problem here was that the world start position of the tool was being compared to a world-projected pixel position, which had vastly different depth values. This has been fixed by storing the start and end position in pixel space and using that for the distance calculation, rather than a world-space position. This would need re-thinking if we wanted the tool to work in perspective viewports, but that was a problem with the old code as well.

Also made the measurement display in more relevant units, rather than just showing raw uu.

This addresses TTP#344821 - CRITICAL: Regression: EDITOR: Ruler gives incorrect measurements.

[CL 2277554 by Andrew Rodham in Main branch]