Merge from shootercert (UE4-Shootergame):
Fix for the network guid issue during seamless travel when we flush guids on the server. The quick fix is to not even send this property to the server, since we don't need to. This has the added benefit of not only fixing the issue, but saves bandwidth. We still need to look into letting guids persist on the server for a bit longer to rully resolve any future conflicts that could happen.

[CL 2278168 by John Pollard in Main branch]