Added csv parser
The csv parser takes a string buffer which it splits into a series of null terminated, unescaped cell strings. In this way it can completely avoid heap allocation (esecially if constructed with a moved string). The parser copes with quoted cell values, and attempts to behave in the same way as Microsoft Excel for interoperability.

[CL 2281417 by Andrew Rodham in Main branch]