Turning off Surface Snapping in the editor viewport now prevents surface alignment during placement.
Previously this would only have been the case if the "Rotate to Surface Normal" option was disabled, but this was hidden and a little confusing.

I've also changed the actor positioning code to take parameter structs that have construction helpers so as to make the code more flexible to the client, whilst keeping the logic in one place.

I've fixed one or two bugs here as well:
	* Fixed vertex snapping helpers not showing up when dragging (they are still hidden when pasting)
	* Fixed point lights not having the same spawn offset as spot and directional lights
	* Fixed newly placed actors not getting their default rotations when not aligning to a surface

This addresses TTP#342860 - Static meshes that are drag and dropped into a world will have their rotation affected by the object the are dragged over
And TTP#341787 - CRITICAL: Regression: EDITOR: Direction light is at a flat angle by default

[CL 2289354 by Andrew Rodham in Main branch]