Source code access now works in SlateViewer and Launcher
Made sure to manually load the appropriate modules for source code access.
The hot reload dependency that VisualStudioSourceCodeAccess had was bringing in the whole editor as a dependency, so this is now restricted to editor builds.
SlateViewer now compiles in CoreUObject as the source code accessor module requires it for its settings.

TTP# 333675 - CHECKIN: SLATE: WidgetReflector code reference hyperlinks don't work anymore

#codereview Nick.Atamas,Frank.Fella

[CL 2292278 by Thomas Sarkanen in Main branch]