Editor: Workaround for color picker themes crash on shutdown
The problem is that the color picker has a global static instance of the color themes widget. By the time that widget is being destroyed as part of static shutdown, Slate and ICU are already unloaded. Ideally, we wouldn't have a global variable here at all. Some further research is needed to understand why it was added in the first place. In the meantime, the variable is being changed to a weak pointer, so it gets destroyed when the last color picker instance gets destroyed.

#CodeReview: jamie.dale, saul.abreu

[CL 2298812 by Max Preussner in Main branch]