Fixed text justification not working correctly when the UI was scaled
The width used for justification is the larger of the "desired size" (as fixed wrapping widths force the desired size to that width), or the "actual size" (in the case that the text is smaller than its containing widget, and the text widget is set to fill width).

The "actual size" wasn't being scaled before being used to offset the text blocks, and also wasn't taking into account the text layout margin. Both of these issues lead to some incorrect offsets being calculated.

Also added a test case for STextBlock justification to the widget gallery in Slate Viewer.

ReviewedBy Nick.Darnell, Nick.Atamas

[CL 2299261 by Jamie Dale in Main branch]