Fixed "flickering" issue caused by SMultiLineEditableText calculating invalid scroll offsets
This only happened when SMultiLineEditableText was being used without scrollbars (as the scrollbars were clamping the offset).

If the content was smaller than the actual size, the ratio calculated would be greater than 1, which would produce bad offsets - the offset is now only calculated when the ratio is <1, otherwise it's just set to 0.

#codereview Nick.Darnell

[CL 2299625 by Jamie Dale in Main branch]