Fixed some layout shuffling when using wrapping with a multi-line editable text
TTP# 347063 - EDITOR: BUG: Multiline editor tries to be larger than it needs to be at first

The scrollbars were always initially made visible, which in the case of wrapping, meant they could appear for a frame and then vanish again once the text had been wrapped. They're now hidden until we have valid wrapping information for the text layout.

This also addresses an issue where the property editor could incorrectly clip off some properties when a multi-line editable text changed size. This was due to the fact that you're required to notify the property editor that is has to refresh when a widget has changed size, and SPropertyEditorText now does this.

ReviewedBy Andrew.Rodham

[CL 2305437 by Jamie Dale in Main branch]