Added the game name to the editor startup analytics
TTP# 346677 - Analytics: Need a discrete new project event so we know when in the project browser or not

The "Editor.ProgramStarted" analytics event now includes a "GameName" attribute containing the contents of FApp::GetGameName(). If this is present and empty, then you know the editor has been opened in "project browser" mode, rather than as a full editor.

#codereview Max.Preussner

[CL 2305551 by Jamie Dale in Main branch]