Adding an updatable texture to easily allow writing to textures every frame.
Added an updatable texture interface, which will be implemented by the main texture types of each slate renderer.
Ensured that it is safe to update/resize the texture from the game or render thread in the case of the RHI renderer.
D3D could be improved if we could write the whole texture at once but the only working method I could find exposes memory buffer with 16 byte alignment for each row.

Fixed non-unity build failure

#codereview Matt.Kuhlenschmidt, John.Pollard

[CL 2306749 by Matthew Griffin in Main branch]