Optimisation for setting unique actor names on multiple objects
Previously the logic was to iterate the entire world each time a duplicate name was found. This was very slow when setting unique actor names for large numbers of objects (especially if they all shared similar names). The function now takes an option set of actor labels to avoid having to iterate the world every time.

[CL 2306780 by Andrew Rodham in Main branch]