Fixed a 1 frame delay in text block size when scaling
We don't know the scale when calling ComputeDesiredSize, so we were trying to cache and use the previous frames scale instead, however that was producing some noticable scaling artefacts on text blocks.

We've "fixed" this by always having ComputeDesiredSize return the text size with a text scale of 1. This means that we might over estimate the amount of space required for the text (as it doesn't scale linearly like the Slate text layout assumes widgets do), but it fixes the size changing after 1 frame (this is how the text block sized used to be calculated before we had text layouts).

#codereview Nick.Atamas, Justin.Sargent

[CL 2308153 by Jamie Dale in Main branch]