Fixed a crash when the TSF IME API tried to get the selected text from SOutputLog
FOutputLogTextLayoutMarshaller::GetText wasn't implemented, so it was always returning an empty string. I've implemented this and verified that it doesn't cause any performance issues for the output log.

I had to change SMultiLineEditableText::LoadText to only do that when it was bound to a delegate, as that was causing a hitch while the output log was updated when you gave it focus.

I also fixed giving the output log focus sometimes jumping to the top of the log (SMultiLineEditableText::OnKeyboardFocusReceived needed the same fix as SMultiLineEditableText::OnKeyboardFocusLost to prevent the cursor jumping into focus).

ReviewedBy Justin.Sargent

[CL 2308259 by Jamie Dale in Main branch]