Fixed auto-wrapped text in the tutorials list never growing again if it was shrunk
TTP# 347884 - EDITOR: TUTORIALS: text does not expand when maximizing the window

The auto-wrapping allowed the text block to shrink when the window was shrunk, however the sizers were set to align their content left horizontally (rather than fill, which is the default) meaning that the text block was never allowed to grow again once it had shrunk.

A better way to achieve this is to set the sizer slots to fill, and then use the text justification on the text block to control the text location - thankfully both of these are the defaults.

#codereview Thomas.Sarkanen

[CL 2313013 by Jamie Dale in Main branch]