Fixed a potential crash when calling SMultiLineEditableText::SetEditableText
TTP# 347830 - EDITOR: CRASH: Possible to get a crash in a multi line text field after resetting it to default

It was possible for the cursor to be left past the end of the newly set text, resulting in a crash the next time the cursor was updated. We now ensure that the cursor is always in a valid position after calling SetEditableText.

ReviewedBy Justin.Sargent

[CL 2313031 by Jamie Dale in Main branch]