Made sure widgets are tagged correctly in tabs
This allows tab content in workflow tabs to properly tag their content.
Viewports now use the metadata from their parent implementation if there is one.

NOTE: TTP listed will not be addressed until a build is promoted with this change and content is edited to fix the issue (i.e. re-tag the viewport).

TTP# 347538 - TUTORIALS: Animation Editor Walkthrough: Multiple Viewports are highlighted when it's trying to showcase the one in persona

[CL 2315651 by Thomas Sarkanen in Main branch]