Allowing transparency in web browser windows and improved viewport resizing
Allow user to chose whether transparency is used for individual browser windows.
When browser viewport is resized or rendered with old size afterwards, made sure that the texture is copied row by row to avoid distortion.
Store size of viewport as an FIntPoint to avoid various casts that would be needed.

[CL 2321219 by Matthew Griffin in Main branch]