Adding external link indicator to tutorial hyperlinks
Slate changes:
Allowed hyperlink decorator and hyperlink run to be more easily derived from by moving some functions/variables to protected.

Added new run & decorator for hyperlinks in tutorials that add an icon to the end of the run if required.
Only use icon-style run if in display mode (not in edit mode).

reviewed by Jamie.Dale

[CL 2322460 by Thomas Sarkanen in Main branch]