Expand Action/Axis mappings when clicking + buttons to add a new one
Setting group expansion at the point an action is added didn't work as the property view is rebuilt before it can be passed on.
Instead replaced existing code to set the expansion state of renamed nodes so that it can be used for any expansion settings that need to be delayed until next tick.

TTP#348688 - NUXF-293 Input binds: Clicking + buttons does not expand the nodes

[CL 2332842 by Matthew Griffin in Main branch]