Fixed a crash caused by routing input events intended for a modal window to a non modal window
The problem was that the slow task dialog has not been written to the hit test grid at the point that we Tick (and subsequently handle an input message for it). Although the message is intended for the slow task window, the widget path generated from the hittest grid said otherwise (returning an SEditorViewport, since that was the last widget rendered to it).

This addresses UE4750 - Crash painting landscape layers during material compilation due to Slate input re-entrancy

#lockdown James.Dickinson

[CL 2343953 by Andrew Rodham in Main branch]