UdpMessaging: Made MulticastSocket creation optional
A multicast socket is not strictly required for the UDP transport to work. For example, when adding devices as static endpoints in the Editor, all communication will be handled through the UnicastSocket alone. This is a temporary workaround for platforms that may not support multicast. A proper fallback to broadcast and then unicast should be implemented in the future.

[CL 2345466 by Max Preussner in Main branch]