Improved the composite font editor layout
It now shows the fonts within a font family in a grid rather than a list. This uses less space, but presents the same information.

The preview text was removed in favor of making the preview viewport handle displaying a preview of all of the fonts that make up the composite font.

[CL 2347166 by Jamie Dale in Main branch]