Added Worlds and Level Blueprints to the scene outliner
This change includes a large rework of the scene outliner to make it more expandable and flexible with what data it supports.
Tree items now support visitation to allow for external logic on custom columns etc without having to type check and cast.
Sorting has been overhauled to be more flexible and efficient. It now supports correct lexical comparison for numbers.

[CL 2354514 by Andrew Rodham in Main branch]