Speculative fix for a crash deleting a material expression node
UE-4821 - Deleting a material node in the latest fortnite build causes a crash without a crash log

I wasn't able to reproduce this crash, but from looking at the logs and the code, it seems this likely failed due to the property handle becoming invalid. I've added some checks to better handle this case.

[CL 2354612 by Jamie Dale in Main branch]