Fixed some jitter when resizing multiline editable text widgets
UE-5236 - Text Jitters in Detail Panel When Creating a Comment in the Material Editor

We have some code which tries to prevent initial layout jitter when auto-wrapping text, and this was checking that the auto-wrap cached size matched the current paint size. This was what was causing the jitter when resizing, as it was constantly hiding the scrollbar when the size had changed. This has now been changed to check that the cached size has been set, as once that's done, we should have also done a valid wrap.

ReviewedBy Andrew.Rodham

[CL 2355769 by Jamie Dale in Main branch]