Fixed FKerningTable causing all sub-fonts to be loaded into memory
FKerningTable used to check whether any font contained within a composite font contained kerning information. This check caused all of the sub-fonts to be loaded into memory, even if they were never used.

This change defers this check until we actually need to get the kerning information for the characters in question.

#codereview Matt.Kuhlenschmidt

[CL 2356012 by Jamie Dale in Main branch]