Made ClearUserFocus default to using EFocusCause::SetDirectly rather than EFocusCause::Cleared
UE-4862 - Content browser search bar is functioning inconsistently.

ClearUserFocus is supposed to be used as a replacement for calling SetUserFocus with a null widget, however it was defaulting to EFocusCause::Cleared rather than EFocusCause::SetDirectly. EFocusCause::Cleared should only be used when you want to unset focus *and* empty the contents of a widget (this is used by text controls when you press escape), so this change meant that some text controls were being cleared out at inappropriate times.

#codereview Chris.Gagnon

[CL 2357031 by Jamie Dale in Main branch]