Fixed incorrect assert in FSlateTextureAtlas
UE-5321 - Regression: Crash using VisualizeTexture from FSlateFontCache

CL# 2350206 made it possible for FSlateTextureAtlas to be created and managed by the render thread as well as the game thread, and the assert in FSlateTextureAtlas::MarkTextureDirty was updated in an attempt to deal with that.

The assert was naive though, as the render thread ID can change if the render thread is suspended and resumed. It now stores the type of thread that owns the FSlateTextureAtlas (rather than its raw thread ID), and then uses that to check that FSlateTextureAtlas::MarkTextureDirty is being called from the correct thread.

[CL 2358609 by Jamie Dale in Main branch]