Text rendering perf. improvements
UETOOL-129 - Investigate SOutputLog performance degradation

This improves the caching of the kerning data used by the Slate font cache, as profiles were showing this to be a hot spot for text rendering.

FKerningTable has been changed to work on a single raw font (FFontData), rather than an entire composite font. This lets FCharacterList only generate a kerning table for those fonts that have kerning, rather than asking each font every time kerning was requested for a character.

This has the side effect of cleaning up the API of FSlateFontCache, as all functions that take a FSlateFontInfo now operate on the composite font as a whole (GetMaxCharacterHeight, GetBaseline), and the functions that take FFontData now operate on a single raw font (GetKerning, HasKerning, GetFontAttributes). This also exposes the GetDefaultFontData and GetFontDataForCharacter so that you can perform more efficient FFontData lookups from outside the FreeType interface.

#codereview Matt.Kuhlenschmidt

[CL 2359866 by Jamie Dale in Main branch]