Text rendering perf. improvements
UETOOL-129 - Investigate SOutputLog performance degradation

This further improves the performance of GetKerning by removing some map lookups.

FCharacterList doesn't really need multiple kerning tables, as each character in a composite font will only ever map to a single raw font for a given FSlateFontInfo. This means we can use a single kerning table, *however* we need to know whether the font a character is using has kerning *without* having to query it from the FreeType interface (as this involves a face lookup, which is relatively slow).

FCharacterList now caches this information in its character entries, so that we can immediately know whether a character will need kerning or not. Additionally, GetKerning can now optionally take a pair of FCharacterEntrys (rather than a pair of TCHARs), which avoids the need to lookup the character entry from the FCharacterList (which is handy when you already have the correct FCharacterEntry to hand, as you do in the case of text batching and measuring).

#codereview Matt.Kuhlenschmidt

[CL 2360190 by Jamie Dale in Main branch]