Sockets: Disabling socket handle inheritance on Windows
Windows enables inheritance for socket inheritance by default, which causes open listener sockets to be passed to client processes. This causes problems in certain scenarios, for example when restarting the Editor to load a new project, or when interfacing with the Android SDK (adb.exe).  Child processes or new instances of the Editor are then not able to bind to the listener ports.

#CodeReview: sam.zamani

[CL 2364217 by Max Preussner in Main branch]