Atlas visualizer improvements, and added a font atlas visualizer
Moved the existing atlas visualizer out of the Slate RHI renderer and instead made the resource manager for each renderer implement the ISlateAtlasProvider interface to provide information about its current texture atlases. FSlateFontCache implements ISlateAtlasProvider for the font atlas viewer.

Improved the atlas visualizer and added support for zooming and panning of the atlas pages. This helps greatly when viewing font atlases as they can be as large as 2048x2048.

#codereview Matt.Kuhlenschmidt

[CL 2371483 by Jamie Dale in Main branch]