Fixed Ctrl+Enter potentially clearing text blocks
UE-5949 - Pressing ctrl+enter in string default variable text field results in text being deleted

Ctrl+Enter was being processed as a confirmation of the text block content, which, depending on the settings of the text block, would sometimes cause all of the text within it to be highlighted. Once this had been done, FTextEditHelper::OnKeyChar was processing the '\\n' character that the Ctrl+Enter combination had caused - this caused any selected to text to be removed, and had the effect of clearing the text block.

FTextEditHelper::OnKeyChar no longer allows '\\n' characters to be processed, as those are handled elsewhere by the editable text boxes that support multiple lines.

#codereview Justin.Sargent

[CL 2383452 by Jamie Dale in Main branch]