Expanding macros with illegally placeable nodes in the context of the current graph will no longer crash.
Assign and temporary variable nodes can now be pasted into any graph with a supporting schema. Drag and drop from palette and context menu will not allow these nodes still.

When dragging and dropping nodes from the palette, if the node cannot be pasted or it is not compatible with the graph, it will not be droppable.

#jira UE-6234 - Expand Node crashes engine

[CL 2383820 by Michael Schoell in Main branch]