Implemented text scaling for composite fonts in Canvas
This was implemented for the bitmap fonts, but hadn't been implemented for composite fonts.

It scales the glyph texture when rendering, rather rendering the glyph with an updated scale. This saves us potentially thrashing the font cache (and having to implement non-uniform text scaling for the font cache), at the cost of text quality.

If you need crisp text, consider scaling the font size rather than using render scaling.

[CL 2388381 by Jamie Dale in Main branch]