Stopped FName converting an ANSICHAR source string to a WIDECHAR string, only to convert it back to ANSICHAR again later
FName::Init could take either an ANSICHAR* or a WIDECHAR*, however the ANSICHAR* variant would convert the string to a WIDECHAR[], and would later convert it back to an ANSICHAR[] because it contains only ANSI characters.

Recent refactoring in FName has made it possible to optimize the ANSICHAR* path, by avoiding the double conversion and just using the ANSICHAR* directly when adding to the FName string table.

ReviewedBy James.Hopkin

[CL 2388399 by Jamie Dale in Main branch]