Allowed UFactory::StaticImportObject to call FactoryCreateNew even when it has a filename set
UDN - bCreateNew=true, and Drag & Drop files with a Factory

This can be useful if your factory needs knowledge of the filename, but doesn't actually do anything with the file data.

In order to try and preserve any existing behavior, the factories are sorted such that these calls to FactoryCreateNew will happen only if there are no factories present that would actually import the file data.

[CL 2388843 by Jamie Dale in Main branch]