Fix for UE-5189 Skel/Vert/Surf module is spawning sprites at emitter origin.
This was being caused by a combination of the daft way the skel/vert/surf module limits particles to an array of valid bone names.
Particles are assigned a vertex/surf index at spawn and this is then checked against the valid bone list. If it is not influenced by a bone on the list it is rejected by setting it's RelativeLife to 1.1.
If a Lifetime module is placed after the skel vert module then this rejection is overwritten and the particle spwans at 0,0,0.
I've added a check for this in the lifetime module to preserve the original RelativeTime if it's already set at > 1.0.
A proper fix to this will require bigger changes to the skel/vert module than I'm willing to delve into now that cascade is being ushered off to the retirement home.

#codereview Olaf.Piesche

[CL 2395019 by Simon Tovey in Main branch]